Move `getInstalledModules` to `iasql.ts`
