+++ src/modules/iasql_functions/index.ts
@@ -64,3 +64,4 @@ export class IasqlFunctions extends ModuleBase {
  * ```
  */
 export const iasqlFunctions = new IasqlFunctions();
+export { getInstalledModules } from './iasql';
